requirements.txt
annotated-types==0.5.0
anyio==3.7.1
bcrypt==4.0.1
brotlipy
cachetools==5.3.1
certifi==2023.7.22
cffi
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
cryptography
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.1.3
fastapi==0.103.0
fastapi_socketio
greenlet
h11==0.14.0
idna==3.4
jmespath
mysqlclient==2.2.0
passlib==1.7.4
pyasn1==0.5.0
pycparser
pydantic==1.10.10
pydantic_core==2.6.3
pyOpenSSL
PySocks
python-dateutil
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
python-socketio
rsa==4.9
s3transfer
six
sniffio==1.3.0
SQLAlchemy~=2.0.25
starlette==0.27.0
typing_extensions
ujson==5.8.0
uritemplate==4.1.1
urllib3
uvicorn==0.23.2
watchfiles==0.20.0
websockets==11.0.3